+++ frontend/src/modules/Metrics/Components/Metrics.tsx
@@ -1,15 +1,9 @@
-import React, { useState } from 'react'
-import Button from '@mui/material/Button'
-import Menu from '@mui/material/Menu'
+import { useState } from 'react'
 import MenuItem from '@mui/material/MenuItem'
 import { ReactComponent as LogoImg } from '@assets/img/lscicon.svg'
-import { KeyboardArrowDown } from '@mui/icons-material'
-import { logOut } from '@fire'
-import { useAuthValue } from '@auth'
 import { useCourseValue } from '@context/CourseContext'
 import { useStudentValue } from '@context/StudentContext'
-import { CSVLink } from 'react-csv'
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft'
+import { AccountMenu } from 'Dashboard/Components/AccountMenu'
 import {
   StyledContainer,
   StyledHeaderMenu,
@@ -21,10 +15,8 @@ import { MetricsTable } from './MetricsTable'
 import { DropdownSelect } from '@core/Components'
 
 export const Metrics = () => {
-  const { user } = useAuthValue()
   const { courses } = useCourseValue()
   const { students } = useStudentValue()
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
 
   //calculate number of unique students who have made requests
   const num_students = {
@@ -79,20 +71,6 @@ export const Metrics = () => {
     num_groups,
   ]
 
-  const csvCourses = courses.map((course) => ({
-    semester: course.roster,
-    course: course.names.join('/'),
-  }))
-
-  const open = Boolean(anchorEl)
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget)
-  }
-  const handleClose = () => {
-    setAnchorEl(null)
-  }
-
   //this can be removed if there is a place to store an objectMap() function
   const localeMap = (obj: { [key: string]: Date } | undefined) => {
     if (!obj) {
@@ -134,15 +112,6 @@ export const Metrics = () => {
         )
       : []
 
-  const [rostorAnchorEl, setRosterAnchorEl] = useState<null | HTMLElement>(null)
-  const openRoster = Boolean(rostorAnchorEl)
-  const handleRosterClick = (event: React.MouseEvent<HTMLElement>) => {
-    setRosterAnchorEl(event.currentTarget)
-  }
-  const handleRosterClose = () => {
-    setRosterAnchorEl(null)
-  }
-
   const [selectedRoster, setSelectedRoster] = useState<string>('FA22')
   const chosenSemesterStudents = csvStudents.filter(
     (e) => e.semester === selectedRoster
@@ -165,12 +134,9 @@ export const Metrics = () => {
     const createdGroups = specificCollegeStudents.filter((student) => {
       return student.groupNumber !== undefined
     })
-    console.log('created Groups')
-    console.log(createdGroups)
     const groups = createdGroups
       .map((s) => s.groupNumber)
       .filter((value, index, self) => self.indexOf(value) === index)
-    console.log(groups)
     return {
       rowName: college,
       students: uniqueStudents.length,
@@ -190,89 +156,10 @@ export const Metrics = () => {
         <Link href="/dashboard" underline="none">
           Dashboard
         </Link>
-        <Button
-          id="logout-button"
-          aria-controls="logout-menu"
-          aria-haspopup="true"
-          aria-expanded={open ? 'true' : undefined}
-          onClick={handleClick}
-          endIcon={<KeyboardArrowDown />}
-          variant="text"
-          disableRipple
-        >
-          {user?.displayName}
-        </Button>
-        <Menu
-          id="logout-menu"
-          anchorEl={anchorEl}
-          open={open}
-          onClose={handleClose}
-          MenuListProps={{
-            'aria-labelledby': 'logout-button',
-          }}
-          anchorOrigin={{
-            vertical: 'bottom',
-            horizontal: 'right',
-          }}
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-        >
-          <CSVLink
-            data={csvCourses.filter((e) => e.semester === selectedRoster)}
-            filename={`export-courses-${Date.now()}`}
-          >
-            <MenuItem>Export CSV (Courses)</MenuItem>
-          </CSVLink>
-          <CSVLink
-            data={csvStudents.filter((e) => e.semester === selectedRoster)}
-            filename={`export-students-${Date.now()}`}
-          >
-            <MenuItem>Export CSV (Students)</MenuItem>
-          </CSVLink>
-          <MenuItem onClick={handleRosterClick}>
-            <ChevronLeftIcon sx={{ color: 'essentials.75', ml: -1 }} />
-            Switch Semester
-          </MenuItem>
-          <MenuItem
-            onClick={() => {
-              handleClose()
-              logOut()
-            }}
-          >
-            Log Out
-          </MenuItem>
-        </Menu>
-        <Menu
-          anchorEl={rostorAnchorEl}
-          open={openRoster}
-          onClick={handleRosterClose}
-          anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'left',
-          }}
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-          sx={{
-            mt: -1.5,
-          }}
-        >
-          <MenuItem onClick={() => setSelectedRoster('SU22')}>
-            Summer 2022
-          </MenuItem>
-          <MenuItem onClick={() => setSelectedRoster('FA22')}>
-            Fall 2022
-          </MenuItem>
-          <MenuItem onClick={() => setSelectedRoster('WI22')}>
-            Winter 2022
-          </MenuItem>
-          <MenuItem onClick={() => setSelectedRoster('SP23')}>
-            Spring 2023
-          </MenuItem>
-        </Menu>
+        <AccountMenu
+          selectedRoster={selectedRoster}
+          setSelectedRoster={setSelectedRoster}
+        ></AccountMenu>
       </StyledHeaderMenu>
       <StyledName>Overall</StyledName>
       <StatGrid stats={stats} />
@@ -306,16 +193,4 @@ export const Metrics = () => {
       />
     </StyledContainer>
   )
-}
-
-interface csvStudentRow {
-  notes: string
-  semester: string
-  dateRequested: string
-  cornellEmail: string
-  name: string
-  college: string
-  year: string
-  course: string
-  groupNumber: string | undefined
 }
